move test processes to conftest
