--- tests/test_processes/wps_test_process.py
@@ -1,44 +1,0 @@
-from pywps import Process, LiteralInput, LiteralOutput
-import logging
-
-LOGGER = logging.getLogger("PYWPS")
-
-
-class TestProcess(Process):
-    """A simple wps process. Outputs the input string."""
-
-    def __init__(self):
-        self.status_percentage_steps = {  # Used to test 'log_handler'
-            "start": 0,
-            "complete": 100,
-        }
-        inputs = [
-            LiteralInput(
-                "string", "String", abstract="Enter string.", data_type="string",
-            )
-        ]
-        outputs = [
-            LiteralOutput(
-                "output",
-                "Output response",
-                abstract="Outputs string entered.",
-                data_type="string",
-            )
-        ]
-
-        super(TestProcess, self).__init__(
-            self._handler,
-            identifier="test_process",
-            title="Test Process",
-            abstract="A simple process that returns the string that was inputted.",
-            inputs=inputs,
-            outputs=outputs,
-            store_supported=True,
-            status_supported=True,
-        )
-
-    @staticmethod
-    def _handler(request, response):
-        LOGGER.info("Begin test process")
-        response.outputs["output"].data = request.inputs["string"][0].data
-        return response
